Improve impl
