+++ cpp/include/kvikio/detail/remote_callback.hpp
@@ -8,7 +8,7 @@
 
 namespace kvikio::detail {
 
-// Defined in cpp/src/remote_handle.cpp. Only used by callback_device_memory(); the host path leaves
+// Defined in cpp/src/remote_handle.cpp. Only used by callback_device_memory(). The host path leaves
 // CallbackContext::bounce_buffer as nullptr.
 class BounceBufferH2D;
 
@@ -17,11 +17,7 @@ class BounceBufferH2D;
  *
  * One instance per byte-range transfer. Captures the destination buffer, the total size requested,
  * and how many bytes have been written so far. The optional `bounce_buffer` is used only by
- * `callback_device_memory()`; the host-memory path leaves it as `nullptr`.
- *
- * @note Lifted out of `remote_handle.cpp`'s anonymous namespace so that the multi-handle backend in
- * `multi_poll_reactor.cpp` (and any future remote backend) can reference the same callbacks and
- * context type across translation units.
+ * `callback_device_memory()`. The host-memory path leaves it as `nullptr`.
  */
 struct CallbackContext {
   char* buf{nullptr};          ///< Output buffer (host memory directly, or staging for device).
@@ -39,7 +35,7 @@ struct CallbackContext {
 };
 
 /**
- * @brief libcurl write callback that copies received bytes directly into a host buffer.
+ * @brief Libcurl write callback that copies received bytes directly into a host buffer.
  *
  * Compatible with `CURLOPT_WRITEFUNCTION`. The `context` argument must point to a `CallbackContext`
  * with the destination host buffer in `ctx->buf` and the expected total size in `ctx->size`. On
@@ -56,7 +52,7 @@ struct CallbackContext {
 std::size_t callback_host_memory(char* data, std::size_t size, std::size_t nmemb, void* context);
 
 /**
- * @brief libcurl write callback that stages received bytes through a pinned host bounce buffer
+ * @brief Libcurl write callback that stages received bytes through a pinned host bounce buffer
  * into device memory.
  *
  * Compatible with `CURLOPT_WRITEFUNCTION`. The `context` must point to a `CallbackContext` whose
@@ -71,4 +67,17 @@ std::size_t callback_host_memory(char* data, std::size_t size, std::size_t nmemb
  */
 std::size_t callback_device_memory(char* data, std::size_t size, std::size_t nmemb, void* context);
 
+/**
+ * @brief Callback for `CURLOPT_WRITEFUNCTION` that copies received data into a `std::string`.
+ *
+ * @param data Received data
+ * @param size Curl internal implementation always sets this parameter to 1
+ * @param num_bytes Number of bytes received
+ * @param userdata Must be cast from `std::string*`
+ * @return The number of bytes consumed by the callback
+ */
+std::size_t callback_get_string_response(char* data,
+                                         std::size_t size,
+                                         std::size_t num_bytes,
+                                         void* userdata);
 }  // namespace kvikio::detail
