--- cpp/include/kvikio/detail/remote_handle.hpp
@@ -1,23 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2025, NVIDIA CORPORATION.
- * SPDX-License-Identifier: Apache-2.0
- */
-#pragma once
-
-#include <cstring>
-
-namespace kvikio::detail {
-/**
- * @brief Callback for `CURLOPT_WRITEFUNCTION` that copies received data into a `std::string`.
- *
- * @param data Received data
- * @param size Curl internal implementation always sets this parameter to 1
- * @param num_bytes Number of bytes received
- * @param userdata Must be cast from `std::string*`
- * @return The number of bytes consumed by the callback
- */
-std::size_t callback_get_string_response(char* data,
-                                         std::size_t size,
-                                         std::size_t num_bytes,
-                                         void* userdata);
-}  // namespace kvikio::detail
